Chore: clean typos and unused code
