+++ src/constants.ts
@@ -6,4 +6,4 @@ export const MAX_PAGE_SIZE = 100;
 
 // iRacing
 export const COOKIE_EXPIRES_IN_MS = 60 * 60 * 1000; // 1 hour
-export const IRACING_URL = "https://members-ng.iracing.com/data";
+export const IRACING_URL = "https://members-ng.iracing.com";
