+++ src/modules/home/ui/views/home-view.tsx
@@ -3,10 +3,8 @@
 import { useTRPC } from "@/trpc/client";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
-import { toast } from "sonner";
-
 export const HomeView = () => {
-  const trpc = useTRPC();
+  // const trpc = useTRPC();
   // const { data } = useSuspenseQuery(
   //   trpc.iracing.getDocumentation.queryOptions(),
   // );
